Move more numeric functions into the numeric module.
